Moved files from interface folder, added comments
