--- VictoryCenter/VictoryCenter.BLL/Interfaces/Search/SearchLogic.cs
@@ -1,8 +1,0 @@
-﻿namespace VictoryCenter.BLL.Interfaces.Search;
-
-// More different types of search can be added here
-public enum SearchLogic
-{
-    Exact,
-    Prefix,
-}
