+++ VictoryCenter/VictoryCenter.UnitTests/ServiceTests/SearchServiceTests.cs
@@ -1,5 +1,6 @@
 ﻿using VictoryCenter.BLL.Interfaces.Search;
 using VictoryCenter.BLL.Services.Search;
+using VictoryCenter.BLL.Services.Search.Helpers;
 using VictoryCenter.DAL.Entities;
 using VictoryCenter.DAL.Enums;
